Implemented post archivation in the backend.
Created new class to archive posts with method that is run every 30 seconds (can be changed later), but it is optimal for demonstration purposes.
Created new status - archived.
Changed the logic of adding new posts - if the status is archived, status is not set to active.
Logic of archiving is to check all posts, skip those that are already archived, compare date and time to current and archive all posts that are in the past.

Also did one small fix in the frontend - 12am is midnight, so changed to 12pm in the dropdown list (for activities between 10am - 12PM and 12PM - 1 pm).